feat: add statistics panel and confirmation modal for farm deletion, enhance layout and responsiveness
